keep track of last pixel in floating value

+++ modular_test/VoltMode.h
@@ -15,7 +15,7 @@ public:
     virtual void exitBrownout();
 
 private:
-    uint16_t _lastPixel;
+    float _lastPixel;
     uint16_t _timeLeft;
 
     void saveToEEPROM();
